oskelly-color-mapping-GL/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
PyYAML>=6.0.0
